Move install info and seccomp out of enabledefault feature (#1866)
* Move all but adp from enabledefault

* Change function name
