--- internal/controller/datadogagent/feature/enabledefault/const.go
@@ -1,12 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-
-package enabledefault
-
-const (
-	DDAgentDaemonSet             = "AGENT_DAEMONSET"
-	DDClusterAgentDeployment     = "CLUSTER_AGENT_DEPLOYMENT"
-	DDDatadogAgentCustomResource = "DATADOGAGENT_CR_NAME"
-)
