--- internal/controller/datadogagent/feature/enabledefault/envvar.go
@@ -1,14 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-
-package enabledefault
-
-const (
-	DDClusterAgentServiceAccountName = "DD_CLUSTER_AGENT_SERVICE_ACCOUNT_NAME"
-
-	// InstallInfoToolVersion is used by the Operator to override the tool
-	// version value in the Agent's install info
-	InstallInfoToolVersion = "DD_TOOL_VERSION"
-)
