+++ internal/controller/datadogagent/feature/livecontainer/feature.go
@@ -65,7 +65,7 @@ func (f *liveContainerFeature) Configure(dda *v2alpha1.DatadogAgent) (reqComp fe
 
 // ManageDependencies allows a feature to manage its dependencies.
 // Feature's dependencies should be added in the store.
-func (f *liveContainerFeature) ManageDependencies(managers feature.ResourceManagers, components feature.RequiredComponents) error {
+func (f *liveContainerFeature) ManageDependencies(managers feature.ResourceManagers) error {
 	return nil
 }
