+++ internal/controller/datadogagent/feature/liveprocess/feature.go
@@ -73,7 +73,7 @@ func (f *liveProcessFeature) Configure(dda *v2alpha1.DatadogAgent) (reqComp feat
 
 // ManageDependencies allows a feature to manage its dependencies.
 // Feature's dependencies should be added in the store.
-func (f *liveProcessFeature) ManageDependencies(managers feature.ResourceManagers, components feature.RequiredComponents) error {
+func (f *liveProcessFeature) ManageDependencies(managers feature.ResourceManagers) error {
 	return nil
 }
