+++ internal/controller/datadogagent/feature/logcollection/feature.go
@@ -82,7 +82,7 @@ func (f *logCollectionFeature) Configure(dda *v2alpha1.DatadogAgent) (reqComp fe
 
 // ManageDependencies allows a feature to manage its dependencies.
 // Feature's dependencies should be added in the store.
-func (f *logCollectionFeature) ManageDependencies(managers feature.ResourceManagers, components feature.RequiredComponents) error {
+func (f *logCollectionFeature) ManageDependencies(managers feature.ResourceManagers) error {
 	return nil
 }
