+++ internal/controller/datadogagent/feature/npm/feature.go
@@ -67,7 +67,7 @@ func (f *npmFeature) Configure(dda *v2alpha1.DatadogAgent) (reqComp feature.Requ
 
 // ManageDependencies allows a feature to manage its dependencies.
 // Feature's dependencies should be added in the store.
-func (f *npmFeature) ManageDependencies(managers feature.ResourceManagers, components feature.RequiredComponents) error {
+func (f *npmFeature) ManageDependencies(managers feature.ResourceManagers) error {
 	return nil
 }
