+++ internal/controller/datadogagent/feature/oomkill/feature.go
@@ -52,7 +52,7 @@ func (f *oomKillFeature) Configure(dda *v2alpha1.DatadogAgent) (reqComp feature.
 
 // ManageDependencies allows a feature to manage its dependencies.
 // Feature's dependencies should be added in the store.
-func (f *oomKillFeature) ManageDependencies(managers feature.ResourceManagers, components feature.RequiredComponents) error {
+func (f *oomKillFeature) ManageDependencies(managers feature.ResourceManagers) error {
 	return nil
 }
