+++ internal/controller/datadogagent/feature/remoteconfig/feature.go
@@ -87,7 +87,7 @@ func (f *rcFeature) Configure(dda *v2alpha1.DatadogAgent) (reqComp feature.Requi
 
 // ManageDependencies allows a feature to manage its dependencies.
 // Feature's dependencies should be added in the store.
-func (f *rcFeature) ManageDependencies(managers feature.ResourceManagers, components feature.RequiredComponents) error {
+func (f *rcFeature) ManageDependencies(managers feature.ResourceManagers) error {
 	return nil
 }
