+++ internal/controller/datadogagent/feature/test/testsuite.go
@@ -123,7 +123,7 @@ func runTest(t *testing.T, tt FeatureTest) {
 	store, depsManager := initDependencies(tt, logger, dda)
 
 	for _, feat := range features {
-		if err := feat.ManageDependencies(depsManager, tt.RequiredComponents); (err != nil) != tt.WantManageDependenciesErr {
+		if err := feat.ManageDependencies(depsManager); (err != nil) != tt.WantManageDependenciesErr {
 			t.Errorf("feature.ManageDependencies() error = %v, wantErr %v", err, tt.WantManageDependenciesErr)
 			return
 		}
