+++ internal/controller/datadogagent/feature/usm/feature.go
@@ -80,7 +80,7 @@ func mergeConfigs(ddaSpec *v2alpha1.DatadogAgentSpec, ddaStatus *v2alpha1.Datado
 
 // ManageDependencies allows a feature to manage its dependencies.
 // Feature's dependencies should be added in the store.
-func (f *usmFeature) ManageDependencies(managers feature.ResourceManagers, components feature.RequiredComponents) error {
+func (f *usmFeature) ManageDependencies(managers feature.ResourceManagers) error {
 	return nil
 }
