+++ internal/controller/datadogagent/finalizer.go
@@ -95,25 +95,22 @@ func (r *Reconciler) finalizeDadV2(reqLogger logr.Logger, obj client.Object) err
 	var errs []error
 
 	// Global dependencies
-	if requiredComponents.ClusterAgent.IsEnabled() {
-		if err := global.ApplyGlobalDependencies(reqLogger, dda, resourceManagers, datadoghqv2alpha1.ClusterAgentComponentName); len(err) > 0 {
-			errs = append(errs, err...)
-		}
+	if err := global.ApplyGlobalDependencies(reqLogger, dda, resourceManagers); len(err) > 0 {
+		errs = append(errs, err...)
 	}
-	if requiredComponents.Agent.IsEnabled() {
-		if err := global.ApplyGlobalDependencies(reqLogger, dda, resourceManagers, datadoghqv2alpha1.NodeAgentComponentName); len(err) > 0 {
-			errs = append(errs, err...)
-		}
+	if err := global.ApplyGlobalComponentDependencies(reqLogger, dda, resourceManagers, datadoghqv2alpha1.ClusterAgentComponentName, requiredComponents.ClusterAgent); len(err) > 0 {
+		errs = append(errs, err...)
 	}
-	if requiredComponents.ClusterChecksRunner.IsEnabled() {
-		if err := global.ApplyGlobalDependencies(reqLogger, dda, resourceManagers, datadoghqv2alpha1.ClusterChecksRunnerComponentName); len(err) > 0 {
-			errs = append(errs, err...)
-		}
+	if err := global.ApplyGlobalComponentDependencies(reqLogger, dda, resourceManagers, datadoghqv2alpha1.NodeAgentComponentName, requiredComponents.Agent); len(err) > 0 {
+		errs = append(errs, err...)
+	}
+	if err := global.ApplyGlobalComponentDependencies(reqLogger, dda, resourceManagers, datadoghqv2alpha1.ClusterChecksRunnerComponentName, requiredComponents.ClusterChecksRunner); len(err) > 0 {
+		errs = append(errs, err...)
 	}
 
 	// Set up dependencies required by enabled features
 	for _, feat := range enabledFeatures {
-		if featErr := feat.ManageDependencies(resourceManagers, requiredComponents); featErr != nil {
+		if featErr := feat.ManageDependencies(resourceManagers); featErr != nil {
 			errs = append(errs, featErr)
 		}
 	}
